+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [#747](https://github.com/spegel-org/spegel/pull/747) Update Go to 1.23.6.
+- [#750](https://github.com/spegel-org/spegel/pull/750) Rename append mirrors to prepend existing.
 
 ### Deprecated